[GameRunner2 Extract Methods] Move getUserRootFolder() and getUserMapsFolder() from GameRunner2 to ClientFileSystemHelper
