+++ src/games/strategy/engine/ClientFileSystemHelper.java
@@ -121,4 +121,29 @@ public final class ClientFileSystemHelper {
     return false;
   }
 
+  public static File getUserRootFolder() {
+    final File userHome = new File(System.getProperties().getProperty("user.home"));
+    // the default
+    File rootDir;
+    if (GameRunner.isMac()) {
+      rootDir = new File(new File(userHome, "Documents"), "triplea");
+    } else {
+      rootDir = new File(userHome, "triplea");
+    }
+    return rootDir;
+  }
+
+  public static File getUserMapsFolder() {
+    final File f = new File(getUserRootFolder(), "maps");
+    if (!f.exists()) {
+      try {
+        f.mkdirs();
+      } catch (final SecurityException e) {
+        e.printStackTrace();
+      }
+    }
+    return f;
+  }
+
+
 }
