+++ src/games/strategy/engine/framework/headlessGameServer/AvailableGames.java
@@ -84,7 +84,7 @@ public class AvailableGames {
   private static List<File> allMapFiles() {
     final List<File> rVal = new ArrayList<File>();
     // prioritize user maps folder over root folder
-    rVal.addAll(safeListFiles(GameRunner2.getUserMapsFolder()));
+    rVal.addAll(safeListFiles(ClientFileSystemHelper.getUserMapsFolder()));
     rVal.addAll(safeListFiles(NewGameChooserModel.getDefaultMapsDir()));
     return rVal;
   }
