+++ src/games/strategy/engine/framework/ui/SaveGameFileChooser.java
@@ -5,6 +5,7 @@ import java.io.File;
 import javax.swing.JFileChooser;
 import javax.swing.filechooser.FileFilter;
 
+import games.strategy.engine.ClientFileSystemHelper;
 import games.strategy.engine.framework.GameRunner2;
 import games.strategy.engine.framework.headlessGameServer.HeadlessGameServer;
 
@@ -14,7 +15,7 @@ public class SaveGameFileChooser extends JFileChooser {
   private static final String AUTOSAVE_2_FILE_NAME = "autosave2.tsvg";
   private static final String AUTOSAVE_ODD_ROUND_FILE_NAME = "autosave_round_odd.tsvg";
   private static final String AUTOSAVE_EVEN_ROUND_FILE_NAME = "autosave_round_even.tsvg";
-  public static final File DEFAULT_DIRECTORY = new File(GameRunner2.getUserRootFolder(), "savedGames");
+  public static final File DEFAULT_DIRECTORY = new File(ClientFileSystemHelper.getUserRootFolder(), "savedGames");
   private static SaveGameFileChooser s_instance;
 
   public enum AUTOSAVE_TYPE {
